stage 0.8

+++ android/src/main/java/com/swmansion/reanimated/NativeProxy.java
@@ -96,7 +96,7 @@ public class NativeProxy {
     mNodesManager.scrollTo(viewTag, x, y, animated);
   }
 
-  private int[] measure(int viewTag) {
+  private float[] measure(int viewTag) {
     return mNodesManager.measure(viewTag);
   }
